A complete DbManager
