+++ src/main/java/server/controller/MainController.java
@@ -1,8 +1,8 @@
 package server.controller;
 
 import com.google.gson.Gson;
+import server.dbmanager.DbManager;
 import server.models.User;
-import server.dbmanager.dbmanager1;
 import server.utility.Digester;
 
 import java.util.Scanner;
@@ -13,13 +13,13 @@ public class MainController {
     private UserController userController;
     private User currentUser;
     private Scanner input;
-    private dbmanager1 dbmanager1;
+    private DbManager dbManager;
     private Digester digester;
 
 
     //The constructor for instantation
     public MainController() {
-        dbmanager1 = new dbmanager1();
+        dbManager = new DbManager();
         digester = new Digester();
 
     }
@@ -27,7 +27,7 @@ public class MainController {
     // Logic behind authorizing user
     public String authUser(String user) {
         User userAuth = new Gson().fromJson(user, User.class);
-        User authorizedUser = dbmanager1.authorizeUser(userAuth.getUsername(), userAuth.getPassword());
+        User authorizedUser = dbManager.authorizeUser(userAuth.getUsername(), userAuth.getPassword());
         String userFound = new Gson().toJson(authorizedUser, User.class);
 
         if (userFound != null) {
@@ -41,7 +41,7 @@ public class MainController {
     public Boolean createUser(String user) {
         User userCreated = new Gson().fromJson(user, User.class);
         userCreated.setPassword(digester.hashWithSalt(userCreated.getPassword()));
-        Boolean ifCreated = dbmanager1.createUser(userCreated);
+        Boolean ifCreated = dbManager.createUser(userCreated);
 
         if(ifCreated) {
             return true;
