+++ src/main/java/server/endpoints/CourseEndpoint.java
@@ -1,19 +1,17 @@
 package server.endpoints;
 
 import com.google.gson.Gson;
-import server.dbmanager.dbmanager1;
+import server.dbmanager.DbManager;
 import server.models.Course;
-//import server.Utility.Digester;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.core.Response;
 import java.util.ArrayList;
-import server.dbmanager.dbmanager3;
 
 @Path("/courses")
 public class CourseEndpoint {
 
-    dbmanager3 dbmanager = new dbmanager3();
+    DbManager dbmanager = new DbManager();
 
     @GET
     public Response loadCourses (){
