+++ src/main/java/server/endpoints/QuestionEndpoint.java
@@ -1,9 +1,6 @@
 package server.endpoints;
 
-import com.google.gson.Gson;
 import server.controller.AdminController;
-import server.dbmanager.dbmanager2;
-import server.models.Question;
 
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
@@ -16,13 +13,12 @@ public class QuestionEndpoint {
 
     @POST
     public Response createQuestion(String questionJson) {
-
         Boolean questionCreated = adminController.createQuestion(questionJson);
 
         return Response
                 .status(200)
                 .type("application/json")
-                .entity("{\"quizCreated\":\"true\"}")
+                .entity("{\"questionCreated\":\"true\"}")
                 .build();
     }
 }
