+++ src/main/java/server/endpoints/UserEndpoint.java
@@ -1,9 +1,8 @@
 package server.endpoints;
 
 import com.google.gson.Gson;
-import server.dbmanager.dbmanager1;
 import server.controller.MainController;
-import server.dbmanager.dbmanager4;
+import server.dbmanager.DbManager;
 import server.models.User;
 import javax.ws.rs.*;
 import javax.ws.rs.core.Response;
@@ -11,8 +10,7 @@ import javax.ws.rs.core.Response;
 @Path("/user")
 public class UserEndpoint {
     //Creating objects of database manager and MainController
-    dbmanager1 dbmanager1 = new dbmanager1();
-    dbmanager4 dbmanager4 = new dbmanager4();
+    DbManager dbManager = new DbManager();
     MainController mainController = new MainController();
     User currentUser = new User();
 
@@ -46,9 +44,8 @@ public class UserEndpoint {
     // User ID as a part of the PATH
     @Path("{id}")
     public Response getUserProfile(@PathParam("id") int id){
-
         //Creates a currentuser from a user ID, which is logged in.
-        currentUser = dbmanager4.getUserProfile(id);
+        currentUser = dbManager.getUserProfile(id);
 
         return Response
                 .status(200)
